refactor(core): extract anchor bootstrap stack (#214)
Made-with: Cursor
